# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Turkey Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Country {
  /** ISO 3166-1 alpha-2 country code (e.g. PK) */
  code: string;
  /** Full country name */
  name: string;
  /** Flag emoji */
  flag: string;
}

export interface SessionInput {
  /** ISO 3166-1 alpha-2 code of the user's passport country */
  passportCountryCode: string;
  /** Anonymous device identifier for server-side history persistence */
  deviceId?: string;
}

export interface ChatSession {
  id: string;
  passportCountryCode: string;
  createdAt: string;
}

export interface MessageInput {
  text: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  sessionId: string;
  role: ChatMessageRole;
  text: string;
  createdAt: string;
}

export type ListChatSessionsParams = {
deviceId: string;
};