# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Turkey Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  /** ISO 3166-1 alpha-2 code of the user's passport country */
  passportCountryCode: string;
  /** Anonymous device identifier for server-side history persistence */
  deviceId?: string;
}